Set up toy cuda dll problem, with a test suite and an example.
Currently attempting to cause device target mismatch between dll and example code, but device pointers are confusing and don't seem to work propperly.

I'm not sure we will be able to achieve the current FLAEM GPU 2 behaviour using a shared object model and may need to go header only / move portions of code to teh header.

One concern is that if the shared object is built for SM35, 50 and 60, and a user compiles for SM 30 or SM70 (or SM 61?) then the final executable may not link. I've not yet managed to demonstrate this however