Investiagted the effects of specifying different / incorrect cuda architectures.
Building the shared lib with just more modern architectures than the hardware does not seem to produce a useful cuda error (or maybe i'm just not checking at the right time).

Building the example with too new an architecture results in an error from cudaMemcpuToSymbol, because the user-specified function is not available for the correct arch.

Specifying a mix of arch's less than or equal to the device used seems ok.